[fix] all: support data from socket under Py3
Socket recv & send work with raw data. Under Py2 it's a `str`. But, under
Py3 it is `bytes`. Hence, simple string constants cannot be successfully
used with socket data under Py3. A "b" prefix to string constant is
required. Under Py2 though `b"a string"` is actually the same thing as
`"a string"` (without "b" prefix).

Moreover, indexing of bytes (`[]`) works differently.
Under Py2: b"abc"[0] == b"a" == "a"
Under Py2: b"abc"[0] == 97 != b"a" and != "a"
But, slicing works similarly:
	b"abc"[:1] == b"a" # under both Py2 & Py3
	b"abc"[-1:] == b"c" # too

This patch reworks code which works with data received from a socket and
data sent to it. So, the code can correctly operate under both Py2 and Py3.

Signed-off-by: Efimov Vasily <real@ispras.ru>