Add support for stubs that don't report features
Currently when pyrsp connects to a stub that does not support feature
reporting, such as kgdb from the Linux kernel, then we get a variety of
crashes due to failure to read from self.feats. Fix this by providing an
empty dictionary when needed and adopting a similar heuristic to gdb to
track the maximum packet size.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>